requirements.txt
flask
google-generativeai
sift-stack-py